+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 * Update react-intl to v5 (UIDATIMP-643)
 * Relocate the file upload area on the Data Import landing page (UIDATIMP-633)
 * Sync with DTO updates. Modifying or updating the SRS MARC record (UIDATIMP-620)
+* Update Datepicker decorator (TextDate component) (UIDATIMP-638)
 * MARC Bib field mapping profile: details for Update Selected fields on View screen (UIDATIMP-614)
 * Add column for Job status and Resequence columns (UIDATIMP-615)
 * Match Profiles: Remove EDIFACT invoice as a match option (UIDATIMP-353)